Moved tabs ui to the /tools/ subfolder
